requirements.txt
numpy
torch
torchvision
pillow
